# apps/api/src/routes/progress.ts
import { zValidator } from "@hono/zod-validator";
import { CompletionPercentageQuerySchema } from "@kubeasy/api-schemas/progress";
import { logger } from "@kubeasy/logger";
import { and, count, eq, inArray, sql } from "drizzle-orm";
import type { Handler } from "hono";
import { Hono } from "hono";
import { nanoid } from "nanoid";
import { db } from "../db/index";
import {
  userProgress,
  userSubmission,
  userXp,
  userXpTransaction,
} from "../db/schema/index";
import { trackChallengeStarted } from "../lib/analytics-server";
import { cacheDel, cacheDelPattern, cached, cacheKey, TTL } from "../lib/cache";
import { getChallenge, listChallenges } from "../lib/registry";
import { requireAuth } from "../middleware/session";

const progress = new Hono();

// GET /progress/completion -- get completion percentage (global or by theme)
progress.get(
  "/completion",
  requireAuth,
  zValidator("query", CompletionPercentageQuerySchema),
  async (c) => {
    const user = c.get("user");
    const userId = user.id;
    const { splitByTheme, themeSlug } = c.req.valid("query");

    const key = cacheKey(`u:${userId}:progress:completion`, {
      splitByTheme: String(splitByTheme),
      themeSlug,
    });

    const data = await cached(key, TTL.USER, async () => {
      if (splitByTheme) {
        // Fetch registry list for theme mapping + user completed slugs from DB
        const [registryList, completedSlugs] = await Promise.all([
          listChallenges(),
          db
            .select({ challengeSlug: userProgress.challengeSlug })
            .from(userProgress)
            .where(
              and(
                eq(userProgress.userId, userId),
                eq(userProgress.status, "completed"),
              ),
            ),
        ]);

        const completedSet = new Set(
          completedSlugs.map((r) => r.challengeSlug),
        );

        // Group by theme
        const themeStats = new Map<
          string,
          { totalCount: number; completedCount: number }
        >();
        for (const ch of registryList) {
          const s = themeStats.get(ch.theme) ?? {
            totalCount: 0,
            completedCount: 0,
          };
          s.totalCount += 1;
          if (completedSet.has(ch.slug)) s.completedCount += 1;
          themeStats.set(ch.theme, s);
        }

        const byTheme = [...themeStats.entries()].map(
          ([themeSlug, { totalCount, completedCount }]) => ({
            themeSlug,
            completedCount,
            totalCount,
            percentageCompleted:
              totalCount > 0
                ? Math.round((completedCount / totalCount) * 100)
                : 0,
          }),
        );

        const totalCount = registryList.length;
        const completedCount = completedSet.size;
        const percentageCompleted =
          totalCount > 0 ? Math.round((completedCount / totalCount) * 100) : 0;

        return { byTheme, completedCount, totalCount, percentageCompleted };
      }

      // Standard mode: single theme or all
      const registryList = await listChallenges();
      const filteredSlugs = themeSlug
        ? registryList
            .filter((ch) => ch.theme === themeSlug)
            .map((ch) => ch.slug)
        : registryList.map((ch) => ch.slug);

      const totalCount = filteredSlugs.length;

      const [completedResult] = await db
        .select({ count: count() })
        .from(userProgress)
        .where(
          and(
            eq(userProgress.userId, userId),
            eq(userProgress.status, "completed"),
            themeSlug && filteredSlugs.length > 0
              ? inArray(userProgress.challengeSlug, filteredSlugs)
              : undefined,
          ),
        );

      const completedCount = completedResult?.count ?? 0;

      return {
        completedCount,
        totalCount,
        percentageCompleted:
          totalCount > 0 ? Math.round((completedCount / totalCount) * 100) : 0,
      };
    });

    return c.json(data);
  },
);

// GET /progress/:slug -- get challenge status for authenticated user
const handleGetStatus: Handler = async (c) => {
  const user = c.get("user");
  const userId = user.id;
  const slug = c.req.param("slug");
  if (!slug) return c.json({ error: "Missing slug" }, 400);

  const data = await cached(
    cacheKey(`u:${userId}:progress:status`, { slug }),
    TTL.USER,
    async () => {
      const [progressRecord] = await db
        .select({
          status: userProgress.status,
          startedAt: userProgress.startedAt,
          completedAt: userProgress.completedAt,
        })
        .from(userProgress)
        .where(
          and(
            eq(userProgress.userId, userId),
            eq(userProgress.challengeSlug, slug),
          ),
        )
        .limit(1);

      return progressRecord
        ? {
            status: progressRecord.status,
            startedAt: progressRecord.startedAt,
            completedAt: progressRecord.completedAt,
          }
        : { status: "not_started" as const };
    },
  );

  return c.json(data);
};

progress.get("/:slug", requireAuth, handleGetStatus);
progress.get("/:slug/status", requireAuth, handleGetStatus);

// POST /progress/:slug/start -- create or update user progress to in_progress
progress.post("/:slug/start", requireAuth, async (c) => {
  const user = c.get("user");
  const userId = user.id;
  const slug = c.req.param("slug");

  const detail = await getChallenge(slug);
  if (!detail) {
    return c.json({ error: "Challenge not found" }, 404);
  }

  const [existingProgress] = await db
    .select({
      id: userProgress.id,
      status: userProgress.status,
      startedAt: userProgress.startedAt,
    })
    .from(userProgress)
    .where(
      and(
        eq(userProgress.userId, userId),
        eq(userProgress.challengeSlug, slug),
      ),
    )
    .limit(1);

  const now = new Date();

  if (existingProgress) {
    if (existingProgress.status === "completed") {
      return c.json({
        status: existingProgress.status,
        startedAt: existingProgress.startedAt,
        message: "Challenge already completed",
      });
    }
    if (existingProgress.status === "not_started") {
      await db
        .update(userProgress)
        .set({ status: "in_progress", startedAt: now })
        .where(eq(userProgress.id, existingProgress.id));
    }
    return c.json({
      status: "in_progress" as const,
      startedAt: existingProgress.startedAt,
    });
  }

  await db.insert(userProgress).values({
    id: nanoid(),
    userId,
    challengeSlug: slug,
    status: "in_progress",
    startedAt: now,
  });

  trackChallengeStarted(userId, slug, detail.title).catch((err) => {
    logger.error("[progress] challenge started tracking failed", {
      userId,
      slug,
      error: String(err),
    });
  });

  Promise.all([
    cacheDel(cacheKey(`u:${userId}:progress:status`, { slug })),
    cacheDelPattern(`cache:u:${userId}:progress:completion:*`),
    cacheDelPattern(`cache:u:${userId}:challenges:list:*`),
  ]).catch((err) => {
    logger.error("[progress/start] cache invalidation failed", {
      userId,
      slug,
      error: String(err),
    });
  });

  return c.json({ status: "in_progress" as const, startedAt: now });
});

// DELETE /progress/:slug/reset
const handleReset: Handler = async (c) => {
  const user = c.get("user");
  const userId = user.id;
  const slug = c.req.param("slug");
  if (!slug) return c.json({ error: "Missing slug" }, 400);

  const detail = await getChallenge(slug);
  if (!detail) {
    return c.json({ error: "Challenge not found" }, 404);
  }

  await Promise.all([
    db
      .delete(userProgress)
      .where(
        and(
          eq(userProgress.userId, userId),
          eq(userProgress.challengeSlug, slug),
        ),
      ),
    db
      .delete(userSubmission)
      .where(
        and(
          eq(userSubmission.userId, userId),
          eq(userSubmission.challengeSlug, slug),
        ),
      ),
    db
      .delete(userXpTransaction)
      .where(
        and(
          eq(userXpTransaction.userId, userId),
          eq(userXpTransaction.challengeSlug, slug),
        ),
      ),
  ]);

  const [xpResult] = await db
    .select({
      totalXp: sql<number>`COALESCE(SUM(${userXpTransaction.xpAmount}), 0)`,
    })
    .from(userXpTransaction)
    .where(eq(userXpTransaction.userId, userId));

  await db
    .update(userXp)
    .set({ totalXp: xpResult?.totalXp ?? 0 })
    .where(eq(userXp.userId, userId));

  cacheDelPattern(`cache:u:${userId}:*`).catch((err) => {
    logger.error("[progress/reset] cache invalidation failed", {
      userId,
      error: String(err),
    });
  });

  return c.json({
    success: true,
    message: "Challenge progress reset successfully",
  });
};

progress.on(["DELETE", "POST"], "/:slug/reset", requireAuth, handleReset);

export { progress };